% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NaturalSplineC}
\alias{NaturalSplineC}
\title{Natural spline method}
\usage{
NaturalSplineC(x)
}
\arguments{
\item{x:}{Attribute vector}

\item{knots:}{Vector of knots (default: all unique values of x)}
}
\value{
: Natural Spline matrix N, each column represents a basis function, each row corresponds to a value of x
}
\description{
Natural spline method
}
