% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscript.R
\name{SmoothingSpline_Univariate}
\alias{SmoothingSpline_Univariate}
\title{implementation of uni-variate spline function}
\usage{
SmoothingSpline_Univariate(x, y, lambda)
}
\description{
implementation of uni-variate spline function
}
